+++ CHANGELOG.md
@@ -9,6 +9,13 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v5.26.1
+
+### Changed
+
+- Improve error message when failing to find classes in namespaces
+- 
+
 ## v5.26.0
 
 ### Added